Restructuring `ratman` crate
Initially I was a little eager to split lots of components
into their own crates, to speed up compliation and add some
separation. But realistically, none of that is really
required. I removed the `routing-core` and `diagnostic` crates
and moved them to their own modules instead.

This way we don't run into problems with circular dependencies.

This commit also changes the way that `Frames` work. First it
re-adds the `trait` abstraction for an `Endpoint`, and also
removes the `broadcast` function, instead relying on `Frame`
to have a `None` `recipient`.

The semantic separation between a `Message` and a `Frame` is
that the latter has sequence numbers, and can thus compose
a Message as a series of smaller chunks.
How the signature should be handled in this case isn't really
clear for now. I guess we might want to replace the `signature`
on Frames alltogether and just have per-frame checksums to make
sure that no delivery errors were introduced.