qaul.net service drafts
This PR adds two services: messaging and files,
which allow people on a qaul network to exchange
text messages (like e-mail) as well as file-share links.

The API of the `messaging` service is flexible enough
to also allow inlining if a payload isn't too big.
Alternatively, the fallback "share link" will be used,
which mirrors the current behaviour of qaul.net (`<2.0.0`)