Implement user_update for Qaul with new UserData
This consolidates the User struct into libqaul::users::User, rather than
having to import users::User as _User. It also seperates out the
UserData so that it's statically impossible to attempt to change the ID
of an existing user, and eliminate a lot of "id == id" checks.

identity::Identity is made public explicitly, since it snuck out in use
users::* and several other crates rely on it.