Adding `Group` to the `Recipient` type in the service API
There's a few questions that come up with this change. Generally, how
low-level or high-level one or the other API should be.

If the service API recipient is aware of groups however, it means that
other services don't have to implement their own. Or that different
services have different understandings of what a group is.

On the other hand, this doesn't allow for any group metadata or
management. It's essentially just a CC, locally managed. A group of
three people would really only exist as a conversation that happens to
send messages to two other parties on each device.

If we had a Group concept in the service API as something that has
members, metadata, is shared across multiple devices, we could then
have services that hook into these groups, which could potentially be
a cool thing to represent in the UI: "This group runs the services:
messaging, filesharing and abolishing capitalism", etc.

For now this isn't implemented at all, but we should think about if
this might be a good idea!