Add `BroadcastMedium` with simulated latency
It is meant to be just one of several simulated media which connect to
`MemMod`s and provide various levels of simulation. `BroadcastMedium`
does no switching and cannot simulate frame loss, but does allow
uniform simulated latency.

It presents an tick-based simulation interface that allows easily
updating the simulation whenever is appropriate to the calling thread.