licenses/agpl-3.0.md: use official Markdown version
Previously this was the plain text version, but had a .md version, so
some programs would probably try to render it as Markdown and end up
with a misrendered document.

Fortunately, the GNU Project has an official Markdown version of the
AGPL that we can use instead.  I obtained this from
<https://www.gnu.org/licenses/agpl.md>.
Message-Id: <20191016180743.24950-3-hi@alyssa.is>