Fundamentally reworking internals of `netmod` abstraction
The idea behind these changes are, that signatures should not be
cryptographic on the lower transfer layer. XxHash was chosen for it's
speed and ease of deployment, comuting checksums for each Frame.

Additionally, the question of how to reassociate a sequence of frames
was addressed in this commit. Instead of just having a Frame, higher
layers in the stack can construct a `Sequence`, which is loaded with
metadata (sender, receiver), and then a series of data payloads (this
assumes properly sliced data already!).

A sequence of Frames consists of a Sequence ID (in this commit set to
all 0), the current checksum and the checksum of the next Frame. The
last Frame contains a `None` for the `next` signature. This way it's
possible to, from the back, rebuild a message sequence, even when
frames arrived out of order.

Payloads are now simply `Vec<u8>`, and there were several smaller
changes around the endpoint API, as well as the Frame internal layout.