Massive API refactoring and core reformatting
This commits changes the way that `libqaul` exports it's primary API
and the types it exposes. Generally, types are now clustered into a
few categories:

- core (root namespace)
- error
- messages (we might shorten this to msg)
- users
- contacts (currently still under users)
- api

The `api` module now only contains the API scope handler types, but
none of the types. While all public types are actually defined in the
api module, they are not exposed there but via the corresponding
module inside libqaul. This means that the API scope becomes more
meaningful again API wise, and export wise, instead of littering it
all into a small export subset.

Other types defined in libqaul should not be public, with a few rare
exceptions. The API is to be considered the main way of interacting
with types.

Additionally this commit changes a bunch of formatting with `cargo
fmt`, that hasn't been run on certain files in weeks. We might want to
add this to our CI at some point, to avoid huge code reformats down
the line.

Signed-off-by: Katharina Fey <kookie@spacekookie.de>