First pass at internal Discovery service structure
There's some problems with the current approach of handling internal
services and I'm not entirely sure what a better approach can even be.
We need to have access to other libqaul internal fields (vial the Qaul
struct), so we need to hand Arc<Qaul> out to every internal
service.

Following are a few approaches we could take in solving this
issue. None of them are _ideal_, although a perfect solution is
probably not possible.

1. Store the Sender behind a Mutex in the Qaul struct
2. Only pass `Sender<DiscCmd>` to whatever service requires it, using
   internal mutability to store it.  This could be done via some kind
   of sevice mesh structure that would allow services to declare
   dependencies on each other and make managing these dependencies
   easier.  At this time I'm not sure what such a mesh would look
   like.
3. Putting Qaul behind a RwLock seems like a bad route to go down

This commit also includes the first real usage of asnyc-std code to
send heartbeat announcements into the network (not active yet).