libqaul: changing messaging APIs to use MsgRef
Now, this seems like a small change but it has a few implications that
we need to be aware of.  Previously ownership of the message was
honded out, and as such, no copy had to ever occur.  This is no longer
the case.  So currently we clone everything no the service
level (messaging in this case), which should really only be done at
the HTTP API level, if ever.

What do we need to change to make this work?  Well, for one we need to
store a copy of the MsgRef so that it con never be cleaned up.  Then,
fields on a TextMessage could simply be an alias to a field present on
the MsgRef.  The only field that really _needs_ to be copied is the
payload, because we need to deserialise it.

I think this is an important issue to solve, but not right now.