libqaul/services: switching to crossbeam channels for registry
This is because the libstd channels can't be made to to send across
thread boundries, even when wrapping them in Send+Sytc wrappers.  The
crossbeam channels have largerly the same API, but provide better
speed, as well a the ability to send acress threads.

This commit changes the channel implementation for the service
registry only.  In the future we should swap all the libstd channel
for crossbeams channels instead.