libqaul/services: ignore usused poll methods
Honestly, I would recommend not using polls for the time being.  Their
behaviour is untested and might not work at all in this model.  We
need to figure out a good way to push to pollers, but also to make
sure that no polls get lost.  The crossbeam channels might help here
because they can be mpmc, but this is not a priority right now.