libqaul/http: porting http-api to latest version of libqaul
This is the second iteration for this commit, without inline comments
with questions (see devel mailing list).  A few caveats: because we
rebased these changes onto master, which has changed, the API remained
broken up to this commit.  It was deemed too much work to port it step
by step.  If you're doing git archeology in the future and hitting
this commit: I'm sorry, the previous 7 (?) are just as relevant.

Otherwise this commit doesn't do anything special.  A few object
semantics changed and stuff needs to be updated.  Formatting and
warnings not yet fixed.