netmod: adding a target ID to endpoint API
We have the problem that when an endpoint implements a one-to-many
mapping, it's not possible for said endpoint to figure out where to
send a frame without having to replicate routing table information in
the driver.

So instead, we introduce the concept of a target in the netmod API,
where an endpoint can keep a local table of availabse endpoints and
what frames have been emitted by them, then exposes this information
to Ratman in a way that doesn't rely on IP addresses or similar
abstractions.  The benifit here is that Ratman can assiciate
Identities with a Interface-Target combo, instead of just the
interface.

Personal note: potentially we want to think about how to express
Floods differently.  Currently they are expressed as -1 because I'm
being paranoid about implementing this stuff on Android in C++ and
having to do a lot of type magic.