Proposing license addition and introducing mailmap
This commit adds a mailmap file to make it easier to associate
contrimutors who use different email addresses or names to commit to
this repository.

The reasoning behind the license change is to allow qaul.net to be
published on the Apple iOS app store.  Because our code repository is
licensed under the AGPL-3.0, we need to use the section 6 additional
permissions section of the AGPL to add this excemption to the license.
This is a change that all current copyright holders (CCed in the patch
mail) need to sign-off on.

Signed-off-by: Katharina Fey <kookie@spacekookie.de>