fixed contributors guide start page links
There were a few wrong links on the contributors start page.
Not all of them failed, as there are still some unremoved documents
on the docs.qaul.net web server. But they all were wrong.

Additionally dependencies.svg file was deleted, as this file is
built automatically by every build of the documentation and
then shows up in git as modified.