libqaul-http: removing from building set
Since we're in the process of moving the entire qaul stack to
async/await, and now the http crate doesn't compile anymore, I think
it's best to just remove the crate from the compilation set for now.
Considering that we might be looking at replacing the json:api with a
more streamlined API approach with websockets anyway, this might be a
good place to make this break.