netmod: changing Endpoint interface to drop &mut requirement (1/4)
This commit changes the endpoint interface, without changing any of
the driver implementations, because more work might be required in
every netmod crate to make this work.  So, to avoid this commit
becoming too big, it's been split up to multiple comits.

The theory behind the change is based on the fact that almost all
interfaces will have to interact with some other kind of runtime or
socket handler that fills a buffer externally.  Because of this there
needs to be some kind of locking mechanism to provide internal
mutability anyway, so the mutable reference becomes less useful.
Additionally the fact that references are immutable makes the
DriverMap easier in the Ratman core, meaning we get to drop some
unsafe code from it.