ratman: removing signature concept
The router is really the wrong place to deal with signatures at the
moment, because we have to way of giving ratman acces to a keystore.
For the moment, we will do the verification in libqaul.  Later on it
would be great if we could have ratman verify messages, once we know
how the frame layout can be sequenced, and signed.