netmod-udp: fixing various bugs in udp socket impl
Previously we bound the udp socket to the same port for all clients,
but this makes it hard (to impossible) to test clients on the same
machine.  Because of this, we now keep track of the peer IP and
Port (in Rust terms a "SocketAddr") in the AddrTable.

This makes the simple 1-to-1 socket send example work.  What's still
non-functioning in this code is the multicasting.  While the Socket
creation ends up joining the multicast ring successfully (aledgedly),
no mulitcast packets are actually ever received.