libqaul/http: read body as string, not json payload
This might be a bug in tide that might be worth looking into further.
For the time being, this fix does the trick since the body can be read
as a string, and then be parsed into a json payload.

It's not ideal, but I'm otherwise confused what the difference between
the two endpoints is supposed to be anyway.  Furthermore, not sure why
rolling back 2ccda79551cbc6734f2b334a1280de9db12192d7 didn't also do
the trick.

Regardless, this works for now!