libqaul: changing Qaul::new() in public API to take a path
The path provided must be valid and writable to create a storage
backend, and will currently panic if it doesn't.  This commit then
also adds tempfile as a dependedcy to the current clients, because
it's easier for testing than having to keep persistent state that
might rapidely change between commits.

This code still doesn't pass testst because most of the userstore
implementation was removed and will have to be fixed in future commits.