libqaul/users: adding key fetch to alexandria backends
Previously this function was handled by in-memory data structures, but
now we just proxy the requests through to alexandria and deal with the
conversion of Record -> types.

In the future maybe we want to look into building a type cache again,
especially for things that are used frequently and don't change often/
if at all.