libqaul/users: implementing store abstraction to insert users
The insert mechanism is more clearly separated between local and
remote users, meaning that some code can be shared more effectively.
This also changes the way that key material is handled, because it's
no longer kept in memory as a type, only in the database and needs to
be retrieved with a query.

In the future we will want to change this again for performance
reasons and keep a type cache in libqaul that is only periodically
updated when changes are made to the underlying data.