alexandria: refactoring tests, documentation and API structure
This is a pretty big commit that touches on a lot of components to
move them around a bit more.  The utils module was way too big and
some of the types have now been pulled out into query.  We might still
wanna move some of the types around but this is gonna be enough to be
useful and much nice to use