alexandria: implementing merging iterator support
The idea here is that a user of the database can send the same query
to different user scopes, and then easily merge the result iterators
together into a single one.  The quality of life improvement might me
small for now, but it will really become useful once iterators have
stable ordering.