alexandria: adding basic tracing logging to public API paths
Tracing allows a user of alexandria to much more easily understand the
program flow of requests, especially because things might be happening
concurrently.  At the moment we only expose very simple info and debug
messages for the API, the inner store and the tag cache.  Logging
should be expanded in the future, and maybe also log non-critical
conditions such as pathological cache misses.