alexandria: let tag cache misses fail more gracefully
While technically correct that a tag cache miss would error, it makes
the program flow in the API much harder to reason about.  Now, when a
cache page is missing, we simply return an empty set to signal that no
paths were matched on the query.

This way we can avoid having to deal with the error we just created,
where the solution would either to let it bubble up (bad), or dealing
with it in the same way as we do now, but one function layer above.