libqaul: refactoring signatures, adding full encryption support
This commit changes the signature layer for ed25519-dalex to
ed25519-salsa20, provided by libnacl.  This has the side-effect of
adding encryption support for all messages dispatched to individual
targets (i.e. non floods).  Signatures are still handled by the
security module, and verified.  Unverifiable messages will be dropped
because it means that a message has been tampered with.

This _should_ take care of all avenues of impersitation, but a proper
security analysis should take place in the future regardless.