async-std: pin version to the 1.5.x versions for now
Updating from 1.5 to 1.6 has introduced a whole slew of bugs and
changes (because many of the things we use are behind the `unstable`
feature flag).  So to avoid having to hunt down more problems now
where none have to exist, we're pinning async-std in all projects to
1.5 (leave the patch unpinned in case there's patches).

If you create a new crate that has to rely on async-std, make sure to
chose the right version to keep the dependency tree as small as
possible.