service/voice: various changes to service internals
Most of the code was already well structured and was not touched.
Nonetheless this commit makes a few separation changes in the internal
layers and brings external API structures in line with how the rest of
the qaul.net ecosystem works.  Following is a list with changes:

- changed subscribers streams to concrete types
- subscriber senders are selectively deallocated via drop
- Manage subscribers via id's that can be referenced
  specifically (vectors have no stable ordering on resizes)
- Moving message deserialisation to msgs module
- Moving call state transitions from lib.rs to protocol.rs
- added a stub to send voice data (receiving should be another
  subscriber object)

In the future we might want to break up the worker.rs file into
smaller functions in protocol.rs because it's a bit long.  But not a
priority right now.

The changes made to the test suite are minor, only to update it to the
new external API.  There's one public api doc test now too.