service/chat: fixing an issue where local room list was empty
The problem was that the room list was populated only by the unread
messages that were divided onto the existing rooms.  A room that had
no unread messages was being dropped.  Ultimately this meant that
after creating a room it would be invisible until there was a response
from the peer that marked it as unread.

This implementation sorts the unread messages into a quickly parsable
map first, then iterates all rooms and simply augments the metadata
where appropriate.