libqaul/rpc: implementing streaming support
The reason we have a stream handler in libqaul-rpc is to avoid having
to duplicate the effort of mapping subscription types to stream
endpoints in the actual rpc implementations (i.e. http, ws, ...),
without then having to teach the libqaul-rpc crate about all possible
transmission formats.

The solution is an inversion of control type that gets passed a
responder type by the calling layer that is able to map responses to
streams, however it wants, and a generic abstraction over
subscriptions coming from libqaul.

Subscriptions are given an ID and inserted into a tracker, that means
that future requests can simply cancel a subscription via the ID that
was generated by libqaul-rpc