ratman: fix broken test
This seq was 2 bytes long with postcard, but with bincode it's 4 bytes
long.  I modified the test to make it so that the length was only
specified in one place, to make it easier to change back if we go back
to postcard in future.  That one check of the length is still there
though as a check that the serialization format doesn't change
unexpectedly.

I also changed some assert!s to assert_eq!s, for nicer error messages.

Signed-off-by: Katharina Fey <kookie@spacekookie.de>