libqaul: fixing serialisation problems around flooded messages
We added the extract_simple_payload function to avoid having to pass
around the user store for messages that we know are not encrypted
anyway.  But in that we created a bug because the function that
_check_ if a message is relevant for it called it on messages that
were addressed.

Now the function chain returns Option<T> all the way through so that
it can conditionally be aborted during the check.