Major structural refactoring (please read)
This commit moves a lot of things around, and it's important to be
aware of this for future merge conflicts!

All network modules have been moved to 'netmods'.  All utility
libraries that didn't have immediate relevancy to the project have
been moved to 'utils'.  All references to previous crates have been
changed in Cargo.toml files, and the Cargo.lock file was updated.
Also, the webgui is now part of 'clients'.

If you get a merge conflict with this commit, make sure that you don't
undo the move-refactoring!