Limit the amount of job runners for tests on CI
We've had issues with libqaul::user_delete not completing when running
in parallel.  So far it's not clear to me why this is happening, and
I have yet to consistently reproduce it.  As a quick fix, limit the
test runs to 1, which seems to improve the rate at which it works.