Various updates to android-support and client/android
This commit changes a lot of things, but in a way represents a working
state interaction between the Kotlin (android) code and the Rust
native code.  Some issues are still present, such as having to
manually std::mem::forget(...) state types, instead of being
automagically forgotten.  But this can be fixed later.

Also, as this change set grows it's becoming obvious that the current
separation of android specifics, and java FFI handles in libqaul isn't
working very well.

When refactoring we should probably either move everything into the
android-support crate, assuming that the APIs of libqaul and similar
probably won't be consumed by other Java applications, or we should
make sure that there is enough tooling extrenally (rust4j for
example), to make it possible to have libqaul/ffi/java bindings that
handle everything from type conversions to java model declarations.