ratman: changing endpoint interface to Arc<impl Endpoint>
This change is needed to allow platform and client specific
trampolines to modify the runtime behaviour of an endpoint, or even
just allow it to be accessed from outside the router.  Once an
endpoint goes into the router, getting it out is somewhat tricky.  It
is much easier however to just make an Arc<T> copy first, then passing
it into the Router.

Ideally there should be a second interface that endpoints need to
implement that allows destruction from externally, making the
remove_ep and get_ep APIs on the router redundant.