webgui: renaming to emberweb
This title is a bit more descriptive and opens up to add different web-style
interfaces in the future.  I'm still not that happy with just leaving it in
the root of the repo, instead of movin it to clients, but we'll see about that
later (technically it's not a client itself, so that wouldn't make sense
either).