netmod-tcp: fixing: issue where dropped connections wouldn't resume
The solution to this problem has been to send explicit hello-responses
from the receiving stream end.  When sending a HELLO a task tries to
read from the sending stream to wait for an ACK.  If this response
times out (TODO: make these values configurable), it will officially
drop the sending connection, meaning that the next send re-initialises
the stream.

With some initial testing this has shown _most_ connection drops to
recover gracefully, but not all.  Further testing is required.