qrpc-sdk: changing the way that capnproto types are used
We won't be using the actual RPC implementation for now.  The system
looks too complicated and does too many things we don't really want it
to do.  The qrpc-sdk can abstract a lot of the ugliness away from the
end-user, and ultimately we just want to pass messages along to the
other side to call some functions.

The only thing the receiving end will have to do is gracefully accept
when fields are missing (and use default values if at all possible),
or report back to a client that it's too old.