fix: block user dialog
The bug here was quite simple: the `pop` callback was being invoked with the parent's `context`.

Flutter uses the navigation stack to know what to `pop`. Whenever we invoke it via `Navigator.pop(context)`, it uses the `context` to derive where the method is being invoked and act accordingly.

By using the wrong `context`, the dialog was not being disposed of, and neither was it returning the value of the confirmation, effectively meaning the user wasn't being blocked.